Artificial-Intelligence/MaskDetection/bin/requirements.txt
Pillow 
numpy
torchvision
matplotlib
pandas 
torch 
bs4
opencv-python 
optuna
scikit-image
torchvision 
pytorchtools
joblib
